# lib/api-zod/src/generated/types/domainList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Domain } from "./domain";

export interface DomainList {
  domains: Domain[];
}