# lib/api-zod/src/generated/types/emailDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmailDetailAttachmentsItem } from "./emailDetailAttachmentsItem";

export interface EmailDetail {
  id: string;
  from: string;
  to: string;
  subject: string;
  date: string;
  htmlBody?: string;
  textBody?: string;
  read: boolean;
  attachments?: EmailDetailAttachmentsItem[];
}