# lib/api-zod/src/generated/types/emailList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmailSummary } from "./emailSummary";

export interface EmailList {
  emails: EmailSummary[];
  total: number;
  page: number;
  totalPages: number;
}