# lib/api-zod/src/generated/types/emailSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EmailSummary {
  id: string;
  from: string;
  subject: string;
  date: string;
  read: boolean;
  preview: string;
}